refactor: 카드 상태 객체 분리
