+++ app/src/androidTest/java/nextstep/payments/AddCardScreenTest.kt
@@ -4,9 +4,9 @@ import androidx.compose.ui.test.assertIsEnabled
 import androidx.compose.ui.test.assertIsNotEnabled
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
-import nextstep.payments.screens.card.CardCompanyState
+import nextstep.payments.screens.card.state.CardCompanyState
+import nextstep.payments.screens.card.state.CardState
 import nextstep.payments.screens.card.update.AddCardScreen
-import nextstep.payments.screens.card.update.CardState
 import nextstep.payments.screens.card.update.UpdateCardUiState
 import org.junit.Rule
 import org.junit.Test
