+++ app/src/main/java/nextstep/payments/components/card/NewPaymentCard.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import nextstep.payments.screens.card.CardCompanyState
+import nextstep.payments.screens.card.state.CardCompanyState
 import nextstep.payments.components.card.elements.CardCompanyName
 import nextstep.payments.components.card.elements.IcChip
 import nextstep.payments.ui.theme.PaymentsTheme
