+++ app/src/main/java/nextstep/payments/components/card/RegisteredPaymentCard.kt
@@ -20,8 +20,8 @@ import nextstep.payments.components.card.elements.IcChip
 import nextstep.payments.components.card.elements.OwnerName
 import nextstep.payments.domain.Card
 import nextstep.payments.domain.CardCompany
-import nextstep.payments.screens.card.CardCompanyState
 import nextstep.payments.screens.card.mapper.toDomain
+import nextstep.payments.screens.card.state.CardCompanyState
 import nextstep.payments.screens.card.mapper.toState
 import nextstep.payments.ui.theme.PaymentsTheme
