+++ app/src/main/java/nextstep/payments/screens/card/update/AddCardScreen.kt
@@ -19,7 +19,8 @@ import androidx.compose.ui.unit.dp
 import nextstep.payments.R
 import nextstep.payments.components.card.EmptyPaymentCard
 import nextstep.payments.components.card.NewPaymentCard
-import nextstep.payments.screens.card.CardCompanyState
+import nextstep.payments.screens.card.state.CardCompanyState
+import nextstep.payments.screens.card.state.CardState
 import nextstep.payments.screens.card.update.components.CardCompanyBottomSheetDialog
 import nextstep.payments.screens.card.update.components.CardInformationInputFields
 import nextstep.payments.screens.card.update.components.UpdateCardTopBar
