+++ app/src/main/java/nextstep/payments/screens/card/update/UpdateCardViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.update
 import nextstep.payments.data.InMemoryPaymentCardsRepository
 import nextstep.payments.domain.Card
 import nextstep.payments.domain.PaymentCardsRepository
-import nextstep.payments.screens.card.CardCompanyState
+import nextstep.payments.screens.card.state.CardCompanyState
 import nextstep.payments.screens.card.mapper.toDomain
 
 class UpdateCardViewModel(
